requirements.txt
# Datapizza AI ecosystem
datapizza-ai==0.0.9
datapizza-ai-clients-anthropic==0.0.4
datapizza-ai-clients-openai==0.0.9
datapizza-ai-vectorstores-qdrant==0.0.7
datapizza-ai-embedders-openai==0.0.5

# Crawler
Scrapy==2.11.2

# HTML processing
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
python-dotenv==1.0.1
tqdm==4.67.1
click==8.1.7

# API Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
